requirements.txt
psycopg2-binary==2.9.10
pyacrcloud==1.0.7
python-decouple==3.8
python-telegram-bot==21.6
sentry-sdk==2.17.0